Lifted up FormNewTask component's state to Planning along with the mutation
